compilers: Put Intel compiler classes into the mixins folder
