+++ mesonbuild/compilers/compilers.py
@@ -27,7 +27,6 @@ from ..mesonlib import (
 from ..envconfig import (
     Properties,
 )
-from .mixins.visualstudio import VisualStudioLikeCompiler
 from .mixins.gnu import GnuCompiler, GnuLikeCompiler
 
 """This file contains the data files of all compilers Meson knows
@@ -1560,102 +1559,3 @@ class ArmclangCompiler:
                 parameter_list[idx] = i[:2] + os.path.normpath(os.path.join(build_dir, i[2:]))
 
         return parameter_list
-
-
-# Tested on linux for ICC 14.0.3, 15.0.6, 16.0.4, 17.0.1, 19.0.0
-class IntelGnuLikeCompiler(GnuLikeCompiler):
-
-    def __init__(self, compiler_type):
-        super().__init__(compiler_type)
-        # As of 19.0.0 ICC doesn't have sanitizer, color, or lto support.
-        #
-        # It does have IPO, which serves much the same purpose as LOT, but
-        # there is an unfortunate rule for using IPO (you can't control the
-        # name of the output file) which break assumptions meson makes
-        self.base_options = ['b_pch', 'b_lundef', 'b_asneeded', 'b_pgo',
-                             'b_coverage', 'b_ndebug', 'b_staticpic', 'b_pie']
-        self.id = 'intel'
-        self.lang_header = 'none'
-
-    def get_optimization_args(self, optimization_level):
-        return clike_optimization_args[optimization_level]
-
-    def get_pch_suffix(self) -> str:
-        return 'pchi'
-
-    def get_pch_use_args(self, pch_dir, header):
-        return ['-pch', '-pch_dir', os.path.join(pch_dir), '-x',
-                self.lang_header, '-include', header, '-x', 'none']
-
-    def get_pch_name(self, header_name):
-        return os.path.basename(header_name) + '.' + self.get_pch_suffix()
-
-    def openmp_flags(self) -> List[str]:
-        if version_compare(self.version, '>=15.0.0'):
-            return ['-qopenmp']
-        else:
-            return ['-openmp']
-
-    def compiles(self, *args, **kwargs):
-        # This covers a case that .get('foo', []) doesn't, that extra_args is
-        # defined and is None
-        extra_args = kwargs.get('extra_args') or []
-        kwargs['extra_args'] = [
-            extra_args,
-            '-diag-error', '10006',  # ignoring unknown option
-            '-diag-error', '10148',  # Option not supported
-            '-diag-error', '10155',  # ignoring argument required
-            '-diag-error', '10156',  # ignoring not argument allowed
-            '-diag-error', '10157',  # Ignoring argument of the wrong type
-            '-diag-error', '10158',  # Argument must be separate. Can be hit by trying an option like -foo-bar=foo when -foo=bar is a valid option but -foo-bar isn't
-            '-diag-error', '1292',   # unknown __attribute__
-        ]
-        return super().compiles(*args, **kwargs)
-
-    def get_profile_generate_args(self):
-        return ['-prof-gen=threadsafe']
-
-    def get_profile_use_args(self):
-        return ['-prof-use']
-
-
-class IntelVisualStudioLikeCompiler(VisualStudioLikeCompiler):
-
-    """Abstractions for ICL, the Intel compiler on Windows."""
-
-    def __init__(self, target: str):
-        super().__init__(target)
-        self.compiler_type = CompilerType.ICC_WIN
-        self.id = 'intel-cl'
-
-    def compile(self, code, *, extra_args=None, **kwargs):
-        # This covers a case that .get('foo', []) doesn't, that extra_args is
-        if kwargs.get('mode', 'compile') != 'link':
-            extra_args = extra_args.copy() if extra_args is not None else []
-            extra_args.extend([
-                '/Qdiag-error:10006',  # ignoring unknown option
-                '/Qdiag-error:10148',  # Option not supported
-                '/Qdiag-error:10155',  # ignoring argument required
-                '/Qdiag-error:10156',  # ignoring not argument allowed
-                '/Qdiag-error:10157',  # Ignoring argument of the wrong type
-                '/Qdiag-error:10158',  # Argument must be separate. Can be hit by trying an option like -foo-bar=foo when -foo=bar is a valid option but -foo-bar isn't
-            ])
-        return super().compile(code, extra_args, **kwargs)
-
-    def get_toolset_version(self) -> Optional[str]:
-        # Avoid circular dependencies....
-        from ..environment import search_version
-
-        # ICL provides a cl.exe that returns the version of MSVC it tries to
-        # emulate, so we'll get the version from that and pass it to the same
-        # function the real MSVC uses to calculate the toolset version.
-        _, _, err = Popen_safe(['cl.exe'])
-        v1, v2, *_ = search_version(err).split('.')
-        version = int(v1 + v2)
-        return self._calculate_toolset_version(version)
-
-    def get_linker_exelist(self):
-        return ['xilink']
-
-    def openmp_flags(self):
-        return ['/Qopenmp']
